factor out config into it's own class
